requirements.txt
numpy
matplotlib
plotly
darts
###

rich
pyrootutils
black

seaborn
typer
python-dotenv

pynvim
debugpy
torchchronos

einops
# jupyter
ipykernel
jupyter
ipywidgets
scikit-learn
captum
aim==3.19.3
torchmetrics
optuna
optuna-dashboard
jwt
datasets


# rebuttal
transformers
#sktime
#ipdb
#tensorboard
#xlrd
#xlwt
#xlutil
#tabulate